Localization

+++ SpeechIT/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render
+from django.urls import reverse
 from django.utils import timezone
 from django.conf import settings
+from django.http import HttpResponseRedirect
 from gtts import gTTS
 from .models import Post
 import hashlib
@@ -23,14 +25,15 @@ def create_mp3(text, lang: str):
 
 
 def get_hash(text: str) -> str:
-    return hashlib.sha512(text.upper().encode()).hexdigest()
+    return hashlib.sha512(text.strip().upper().encode()).hexdigest()
 
 
 def del_old():
     posts = Post.objects.all()
     for post in posts:
         if post.need_delete():
-            os.remove(post.path)
+            if os.path.isfile(post.path):
+                os.remove(post.path)
             post.delete()
 
 
@@ -39,19 +42,23 @@ def index(request):
     return render(request, 'SpeechIT/homepage.html', {'path': request.path})
 
 
-def to_speech(request, text_data):
+def to_speech(request):
     text_data = request.POST['qtext']
     text_lang = request.POST['lang']
     hash_digit = get_hash(text_data)
     ip = get_ip(request)
-    from_db = Post.objects.update_or_create(hash_str=hash_digit, defaults={'text': text_data, 'lang': text_lang,
-                                                                           'ip': ip, 'date': timezone.now(),
-                                                                           'path': os.path.join(settings.MEDIA_ROOT,
-                                                                                                f'{hash_digit}.mp3')})
+    Post.objects.update_or_create(hash_str=hash_digit, defaults={'text': text_data, 'lang': text_lang,
+                                                                 'ip': ip, 'date': timezone.now(),
+                                                                 'path': os.path.join(settings.MEDIA_ROOT,
+                                                                                      f'{hash_digit}.mp3')})
+    return HttpResponseRedirect(reverse('SpeechIT:audio_page', args=(hash_digit,)))
 
-    if not os.path.isfile(from_db[0].path):
-        file = gTTS(text=text_data, lang=text_lang)
-        file.save(from_db[0].path)
+
+def audio_page(request, hash_):
+    from_db = Post.objects.get(hash_str=hash_)
+    if not os.path.isfile(from_db.path):
+        file = gTTS(text=from_db.text, lang=from_db.lang)
+        file.save(from_db.path)
 
     return render(request, 'SpeechIT/download.html',
-                  {'db': from_db[0], 'media': f"{settings.MEDIA_URL}{hash_digit}.mp3"})
+                  {'db': from_db, 'media': f"{settings.MEDIA_URL}{hash_}.mp3"})
